Recorder: Fix UI and `RecorderViewModel`
This commit is integrated with the last and changes the `RecorderViewModel`
to manage the service in tandem with the main activity.

In this commit the UI is also changed in different ways.

+++ recorder/src/main/java/com/handmonitor/recorder/data/Action.kt
@@ -12,5 +12,5 @@ class Action {
         Housework
     }
 
-    class TimeRange(val remaining: Int, val total: Int)
+    class TimeRange(val remaining: Float, val total: Float)
 }
